@@ -5,209 +5,29 @@ struct ARMv7Context;
 
 namespace vm
 {
-	template<typename T, int lvl = 1, typename AT = u32>
+	template<typename T, typename AT = u32>
 	struct _ptr_base
 	{
 		AT m_addr;
 
-		typedef typename std::remove_cv<T>::type type;
-		static const u32 address_size = sizeof(AT);
+		using type = T;
 
-		_ptr_base operator++ (int)
-		{
-			AT result = m_addr;
-			m_addr += address_size;
-			return make(result);
-		}
-
-		_ptr_base& operator++ ()
-		{
-			m_addr += address_size;
-			return *this;
-		}
-
-		_ptr_base operator-- (int)
-		{
-			AT result = m_addr;
-			m_addr -= address_size;
-			return make(result);
-		}
-
-		_ptr_base& operator-- ()
-		{
-			m_addr -= address_size;
-			return *this;
-		}
-
-		_ptr_base& operator += (AT count)
-		{
-			m_addr += count * address_size;
-			return *this;
-		}
-
-		_ptr_base& operator -= (AT count)
-		{
-			m_addr -= count * address_size;
-			return *this;
-		}
-
-		_ptr_base operator + (typename	remove_be_t<AT>::type count) const { return make(m_addr + count * address_size); }
-		_ptr_base operator + (typename		to_be_t<AT>::type count) const { return make(m_addr + count * address_size); }
-		_ptr_base operator - (typename	remove_be_t<AT>::type count) const { return make(m_addr - count * address_size); }
-		_ptr_base operator - (typename		to_be_t<AT>::type count) const { return make(m_addr - count * address_size); }
-
-		force_inline bool operator <(const _ptr_base& right) const { return m_addr < right.m_addr; }
-		force_inline bool operator <=(const _ptr_base& right) const { return m_addr <= right.m_addr; }
-		force_inline bool operator >(const _ptr_base& right) const { return m_addr > right.m_addr; }
-		force_inline bool operator >=(const _ptr_base& right) const { return m_addr >= right.m_addr; }
-		force_inline bool operator ==(const _ptr_base& right) const { return m_addr == right.m_addr; }
-		force_inline bool operator !=(const _ptr_base& right) const { return m_addr != right.m_addr; }
-		force_inline bool operator ==(const nullptr_t& right) const { return m_addr == 0; }
-		force_inline bool operator !=(const nullptr_t& right) const { return m_addr != 0; }
-		explicit operator bool() const { return m_addr != 0; }
-
-		force_inline _ptr_base<T, lvl - 1, typename std::conditional<is_be_t<T>::value, typename to_be_t<AT>::type, AT>::type> operator *() const
-		{
-			AT addr = convert_le_be<AT>(read64(convert_le_be<u32>(m_addr)));
-			return (_ptr_base<T, lvl - 1, typename std::conditional<is_be_t<T>::value, typename to_be_t<AT>::type, AT>::type>&)addr;
-		}
-
-		force_inline _ptr_base<T, lvl - 1, typename std::conditional<is_be_t<T>::value, typename to_be_t<AT>::type, AT>::type> operator [](AT index) const
-		{
-			AT addr = convert_le_be<AT>(read64(convert_le_be<u32>(m_addr + 8 * index)));
-			return (_ptr_base<T, lvl - 1, typename std::conditional<is_be_t<T>::value, typename to_be_t<AT>::type, AT>::type>&)addr;
-		}
-
-		template<typename AT2>
-		operator _ptr_base<T, lvl, AT2>() const
-		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<T, lvl, AT2>&>(addr);
-		}
-		
-		AT addr() const
-		{
-			return m_addr;
-		}
-
-		template<typename U>
-		void set(U&& value)
-		{
-			m_addr = convert_le_be<AT>(value);
-		}
-
-		static _ptr_base make(const AT& addr)
-		{
-			return reinterpret_cast<const _ptr_base&>(addr);
-		}
-
-		_ptr_base& operator = (const _ptr_base& right) = default;
-	};
-	
-	template<typename T, typename AT>
-	struct _ptr_base<T, 1, AT>
-	{
-		AT m_addr;
-		
 		static_assert(!std::is_pointer<T>::value, "vm::_ptr_base<> error: invalid type (pointer)");
 		static_assert(!std::is_reference<T>::value, "vm::_ptr_base<> error: invalid type (reference)");
-		typedef typename std::remove_cv<T>::type type;
 
-		force_inline static const u32 data_size()
-		{
-			return convert_le_be<AT>(sizeof(T));
-		}
-
-		force_inline T* const operator -> () const
-		{
-			return vm::get_ptr<T>(vm::cast(m_addr));
-		}
-
-		_ptr_base operator++ (int)
-		{
-			AT result = m_addr;
-			m_addr += data_size();
-			return make(result);
-		}
-
-		_ptr_base& operator++ ()
-		{
-			m_addr += data_size();
-			return *this;
-		}
-
-		_ptr_base operator-- (int)
-		{
-			AT result = m_addr;
-			m_addr -= data_size();
-			return make(result);
-		}
-
-		_ptr_base& operator-- ()
-		{
-			m_addr -= data_size();
-			return *this;
-		}
-
-		_ptr_base& operator += (AT count)
-		{
-			m_addr += count * data_size();
-			return *this;
-		}
-
-		_ptr_base& operator -= (AT count)
-		{
-			m_addr -= count * data_size();
-			return *this;
-		}
-
-		_ptr_base operator + (typename	remove_be_t<AT>::type count) const { return make(convert_le_be<AT>(convert_le_be<decltype(count)>(m_addr) + count * convert_le_be<decltype(count)>(data_size()))); }
-		_ptr_base operator + (typename		to_be_t<AT>::type count) const { return make(convert_le_be<AT>(convert_le_be<decltype(count)>(m_addr) + count * convert_le_be<decltype(count)>(data_size()))); }
-		_ptr_base operator - (typename	remove_be_t<AT>::type count) const { return make(convert_le_be<AT>(convert_le_be<decltype(count)>(m_addr) - count * convert_le_be<decltype(count)>(data_size()))); }
-		_ptr_base operator - (typename		to_be_t<AT>::type count) const { return make(convert_le_be<AT>(convert_le_be<decltype(count)>(m_addr) - count * convert_le_be<decltype(count)>(data_size()))); }
-
-		force_inline T& operator *() const
-		{
-			return vm::get_ref<T>(vm::cast(m_addr));
-		}
-
-		force_inline T& operator [](typename remove_be_t<AT>::type index) const
-		{
-			return vm::get_ref<T>(vm::cast(m_addr + data_size() * index));
-		}
-
-		force_inline T& operator [](typename to_be_t<AT>::forced_type index) const
-		{
-			return vm::get_ref<T>(vm::cast(m_addr + data_size() * index));
-		}
-
-		force_inline bool operator <(const _ptr_base& right) const { return m_addr < right.m_addr; }
-		force_inline bool operator <=(const _ptr_base& right) const { return m_addr <= right.m_addr; }
-		force_inline bool operator >(const _ptr_base& right) const { return m_addr > right.m_addr; }
-		force_inline bool operator >=(const _ptr_base& right) const { return m_addr >= right.m_addr; }
-		force_inline bool operator ==(const _ptr_base& right) const { return m_addr == right.m_addr; }
-		force_inline bool operator !=(const _ptr_base& right) const { return m_addr != right.m_addr; }
-		force_inline bool operator ==(const nullptr_t& right) const { return m_addr == 0; }
-		force_inline bool operator !=(const nullptr_t& right) const { return m_addr != 0; }
-		explicit operator bool() const { return m_addr != 0; }
-		explicit operator T*() const { return get_ptr(); }
-		
 		AT addr() const
 		{
 			return m_addr;
 		}
-
-		template<typename U>
-		void set(U&& value)
+		
+		template<typename CT> std::enable_if_t<std::is_assignable<AT&, CT>::value> set(const CT& value)
 		{
-			m_addr = convert_le_be<AT>(value);
+			m_addr = value;
 		}
 
-		template<typename AT2>
-		operator _ptr_base<T, 1, AT2>() const
+		template<typename AT2 = AT> static _ptr_base make(const AT2& addr)
 		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<T, 1, AT2>&>(addr);
+			return{ convert_le_be<AT>(addr) };
 		}
 
 		T* get_ptr() const
@@ -219,140 +39,59 @@ namespace vm
 		{
 			return vm::priv_ptr<T>(vm::cast(m_addr));
 		}
-		
-		static const _ptr_base make(const AT& addr)
-		{
-			return reinterpret_cast<const _ptr_base&>(addr);
-		}
 
-		_ptr_base& operator = (const _ptr_base& right) = default;
-	};
-
-	template<typename AT>
-	struct _ptr_base<void, 1, AT>
-	{
-		AT m_addr;
-		
-		AT addr() const
-		{
-			return m_addr;
-		}
-
-		template<typename U>
-		void set(U&& value)
-		{
-			m_addr = convert_le_be<AT>(value);
-		}
-
-		void* get_ptr() const
-		{
-			return vm::get_ptr<void>(vm::cast(m_addr));
-		}
-
-		void* priv_ptr() const
-		{
-			return vm::priv_ptr<void>(vm::cast(m_addr));
-		}
-
-		explicit operator void*() const
+		T* operator ->() const
 		{
 			return get_ptr();
 		}
 
-		force_inline bool operator <(const _ptr_base& right) const { return m_addr < right.m_addr; }
-		force_inline bool operator <=(const _ptr_base& right) const { return m_addr <= right.m_addr; }
-		force_inline bool operator >(const _ptr_base& right) const { return m_addr > right.m_addr; }
-		force_inline bool operator >=(const _ptr_base& right) const { return m_addr >= right.m_addr; }
-		force_inline bool operator ==(const _ptr_base& right) const { return m_addr == right.m_addr; }
-		force_inline bool operator !=(const _ptr_base& right) const { return m_addr != right.m_addr; }
-		force_inline bool operator ==(const nullptr_t& right) const { return m_addr == 0; }
-		force_inline bool operator !=(const nullptr_t& right) const { return m_addr != 0; }
-		explicit operator bool() const { return m_addr != 0; }
-
-		template<typename AT2>
-		operator _ptr_base<void, 1, AT2>() const
+		std::add_lvalue_reference_t<T> operator [](u32 index) const
 		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<void, 1, AT2>&>(addr);
+			static_assert(!std::is_void<T>::value, "vm::_ptr_base<> error: operator[] is not available for void pointers");
+
+			return vm::get_ref<T>(vm::cast(m_addr + sizeof32(T) * index));
 		}
 
-		template<typename AT2>
-		operator _ptr_base<const void, 1, AT2>() const
+		// enable only the conversions which are originally possible between pointer types
+		template<typename T2, typename AT2, typename dummy = std::enable_if_t<std::is_convertible<T*, T2*>::value>> operator _ptr_base<T2, AT2>() const
 		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<const void, 1, AT2>&>(addr);
+			return{ convert_le_be<AT2>(vm::cast(m_addr)) };
 		}
 
-		static const _ptr_base make(const AT& addr)
-		{
-			return reinterpret_cast<const _ptr_base&>(addr);
-		}
-
-		_ptr_base& operator = (const _ptr_base& right) = default;
-	};
-
-	template<typename AT>
-	struct _ptr_base<const void, 1, AT>
-	{
-		AT m_addr;
-
-		AT addr() const
-		{
-			return m_addr;
-		}
-
-		template<typename U>
-		void set(U&& value)
-		{
-			m_addr = convert_le_be<AT>(value);
-		}
-
-		const void* get_ptr() const
-		{
-			return vm::get_ptr<const void>(vm::cast(m_addr));
-		}
-
-		const void* priv_ptr() const
-		{
-			return vm::priv_ptr<const void>(vm::cast(m_addr));
-		}
-
-		explicit operator const void*() const
+		template<typename T2, typename dummy = std::enable_if_t<std::is_convertible<T*, T2*>::value>> explicit operator T2*() const
 		{
 			return get_ptr();
 		}
 
-		force_inline bool operator <(const _ptr_base& right) const { return m_addr < right.m_addr; }
-		force_inline bool operator <=(const _ptr_base& right) const { return m_addr <= right.m_addr; }
-		force_inline bool operator >(const _ptr_base& right) const { return m_addr > right.m_addr; }
-		force_inline bool operator >=(const _ptr_base& right) const { return m_addr >= right.m_addr; }
-		force_inline bool operator ==(const _ptr_base& right) const { return m_addr == right.m_addr; }
-		force_inline bool operator !=(const _ptr_base& right) const { return m_addr != right.m_addr; }
-		force_inline bool operator ==(const nullptr_t& right) const { return m_addr == 0; }
-		force_inline bool operator !=(const nullptr_t& right) const { return m_addr != 0; }
-		explicit operator bool() const { return m_addr != 0; }
-
-		template<typename AT2>
-		operator _ptr_base<const void, 1, AT2>() const
+		explicit operator bool() const
 		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<const void, 1, AT2>&>(addr);
+			return m_addr != 0;
 		}
 
-		static const _ptr_base make(const AT& addr)
-		{
-			return reinterpret_cast<const _ptr_base&>(addr);
-		}
-
-		_ptr_base& operator = (const _ptr_base& right) = default;
+		_ptr_base& operator =(const _ptr_base&) = default;
 	};
 
 	template<typename AT, typename RT, typename ...T>
-	struct _ptr_base<RT(T...), 1, AT>
+	struct _ptr_base<RT(T...), AT>
 	{
 		AT m_addr;
 
-		typedef RT(type)(T...);
+		using type = func_def<RT(T...)>;
+
+		AT addr() const
+		{
+			return m_addr;
+		}
+
+		template<typename CT> std::enable_if_t<std::is_assignable<AT&, CT>::value> set(const CT& value)
+		{
+			m_addr = value;
+		}
+
+		template<typename AT2 = AT> static _ptr_base make(const AT2& addr)
+		{
+			return{ convert_le_be<AT>(addr) };
+		}
 
 		// defined in CB_FUNC.h, call using specified PPU thread context
 		RT operator()(PPUThread& CPU, T... args) const;
@@ -363,50 +102,33 @@ namespace vm
 		// defined in CB_FUNC.h, call using current PPU thread context
 		RT operator()(T... args) const;
 
-		AT addr() const
+		// conversion to function object
+		operator std::function<type>() const
 		{
-			return m_addr;
+			const u32 addr = vm::cast(m_addr);
+
+			return [addr](T... args) -> RT
+			{
+				return _ptr_base<RT(T...)>{ addr }(args...);
+			};
 		}
 
-		template<typename U>
-		void set(U&& value)
+		// conversion to another function pointer
+		template<typename AT2> operator _ptr_base<type, AT2>() const
 		{
-			m_addr = convert_le_be<AT>(value);
+			return{ convert_le_be<AT2>(vm::cast(m_addr)) };
 		}
 
-		force_inline bool operator <(const _ptr_base& right) const { return m_addr < right.m_addr; }
-		force_inline bool operator <=(const _ptr_base& right) const { return m_addr <= right.m_addr; }
-		force_inline bool operator >(const _ptr_base& right) const { return m_addr > right.m_addr; }
-		force_inline bool operator >=(const _ptr_base& right) const { return m_addr >= right.m_addr; }
-		force_inline bool operator ==(const _ptr_base& right) const { return m_addr == right.m_addr; }
-		force_inline bool operator !=(const _ptr_base& right) const { return m_addr != right.m_addr; }
-		force_inline bool operator ==(const nullptr_t& right) const { return m_addr == 0; }
-		force_inline bool operator !=(const nullptr_t& right) const { return m_addr != 0; }
-		explicit operator bool() const { return m_addr != 0; }
-
-		template<typename AT2>
-		operator _ptr_base<type, 1, AT2>() const
+		explicit operator bool() const
 		{
-			AT2 addr = convert_le_be<AT2>(m_addr);
-			return reinterpret_cast<_ptr_base<type, 1, AT2>&>(addr);
+			return m_addr != 0;
 		}
 
-		static const _ptr_base make(const AT& addr)
-		{
-			return reinterpret_cast<const _ptr_base&>(addr);
-		}
-
-		operator const std::function<type>() const
-		{
-			const AT addr = convert_le_be<AT>(m_addr);
-			return [addr](T... args) -> RT { return make(addr)(args...); };
-		}
-
-		_ptr_base& operator = (const _ptr_base& right) = default;
+		_ptr_base& operator =(const _ptr_base&) = default;
 	};
 
 	template<typename AT, typename RT, typename ...T>
-	struct _ptr_base<RT(*)(T...), 1, AT>
+	struct _ptr_base<RT(*)(T...), AT>
 	{
 		AT m_addr;
 
@@ -414,39 +136,51 @@ namespace vm
 	};
 
 	// Native endianness pointer to LE data
-	template<typename T, int lvl = 1, typename AT = u32> using ptrl = _ptr_base<typename to_le_t<T>::type, lvl, AT>;
+	template<typename T, typename AT = u32> using ptrl = _ptr_base<to_le_t<T>, AT>;
 
 	// Native endianness pointer to BE data
-	template<typename T, int lvl = 1, typename AT = u32> using ptrb = _ptr_base<typename to_be_t<T>::type, lvl, AT>;
+	template<typename T, typename AT = u32> using ptrb = _ptr_base<to_be_t<T>, AT>;
 
 	// BE pointer to LE data
-	template<typename T, int lvl = 1, typename AT = u32> using bptrl = _ptr_base<typename to_le_t<T>::type, lvl, typename to_be_t<AT>::type>;
+	template<typename T, typename AT = u32> using bptrl = _ptr_base<to_le_t<T>, to_be_t<AT>>;
 
 	// BE pointer to BE data
-	template<typename T, int lvl = 1, typename AT = u32> using bptrb = _ptr_base<typename to_be_t<T>::type, lvl, typename to_be_t<AT>::type>;
+	template<typename T, typename AT = u32> using bptrb = _ptr_base<to_be_t<T>, to_be_t<AT>>;
 
 	// LE pointer to LE data
-	template<typename T, int lvl = 1, typename AT = u32> using lptrl = _ptr_base<typename to_le_t<T>::type, lvl, typename to_le_t<AT>::type>;
+	template<typename T, typename AT = u32> using lptrl = _ptr_base<to_le_t<T>, to_le_t<AT>>;
 
 	// LE pointer to BE data
-	template<typename T, int lvl = 1, typename AT = u32> using lptrb = _ptr_base<typename to_be_t<T>::type, lvl, typename to_le_t<AT>::type>;
+	template<typename T, typename AT = u32> using lptrb = _ptr_base<to_be_t<T>, to_le_t<AT>>;
 
 	namespace ps3
 	{
 		// default pointer for PS3 HLE functions (Native endianness pointer to BE data)
-		template<typename T, int lvl = 1, typename AT = u32> using ptr = ptrb<T, lvl, AT>;
+		template<typename T, typename AT = u32> using ptr = ptrb<T, AT>;
+
+		// default pointer to pointer for PS3 HLE functions (Native endianness pointer to BE pointer to BE data)
+		template<typename T, typename AT = u32, typename AT2 = u32> using pptr = ptr<ptr<T, AT2>, AT>;
 
 		// default pointer for PS3 HLE structures (BE pointer to BE data)
-		template<typename T, int lvl = 1, typename AT = u32> using bptr = bptrb<T, lvl, AT>;
+		template<typename T, typename AT = u32> using bptr = bptrb<T, AT>;
+
+		// default pointer to pointer for PS3 HLE structures (BE pointer to BE pointer to BE data)
+		template<typename T, typename AT = u32, typename AT2 = u32> using bpptr = bptr<ptr<T, AT2>, AT>;
 	}
 
 	namespace psv
 	{
 		// default pointer for PSV HLE functions (Native endianness pointer to LE data)
-		template<typename T, int lvl = 1, typename AT = u32> using ptr = ptrl<T, lvl, AT>;
+		template<typename T> using ptr = ptrl<T>;
+
+		// default pointer to pointer for PSV HLE functions (Native endianness pointer to LE pointer to LE data)
+		template<typename T> using pptr = ptr<ptr<T>>;
 
 		// default pointer for PSV HLE structures (LE pointer to LE data)
-		template<typename T, int lvl = 1, typename AT = u32> using lptr = lptrl<T, lvl, AT>;
+		template<typename T> using lptr = lptrl<T>;
+
+		// default pointer to pointer for PSV HLE structures (LE pointer to LE pointer to LE data)
+		template<typename T> using lpptr = lptr<ptr<T>>;
 	}
 
 	// PS3 emulation is main now, so lets it be as default
@@ -454,27 +188,198 @@ namespace vm
 
 	struct null_t
 	{
-		template<typename T, int lvl, typename AT> operator _ptr_base<T, lvl, AT>() const
+		template<typename T, typename AT> operator _ptr_base<T, AT>() const
 		{
-			const std::array<AT, 1> value = {};
-			return _ptr_base<T, lvl, AT>::make(value[0]);
+			return{};
 		}
 	};
 
 	// vm::null is convertible to any vm::ptr type as null pointer in virtual memory
 	static null_t null;
+
+	// helper SFINAE type for vm::_ptr_base comparison operators (enables comparison between equal types and between any type and void*)
+	template<typename T1, typename T2, typename RT> using if_comparable_t = std::enable_if_t<
+		std::is_void<T1>::value ||
+		std::is_void<T2>::value ||
+		std::is_same<std::remove_cv_t<T1>, std::remove_cv_t<T2>>::value,
+		RT>;
+
+	// helper SFINAE type for vm::_ptr_base pointer arithmetic operators and indirection (disabled for void and function pointers)
+	template<typename T, typename RT> using if_arithmetical_t = std::enable_if_t<
+		!std::is_void<T>::value && !std::is_function<T>::value,
+		RT>;
 }
+
+// unary plus operator for vm::_ptr_base (always available)
+template<typename T, typename AT> inline vm::_ptr_base<T, AT> operator +(const vm::_ptr_base<T, AT>& ptr)
+{
+	return ptr;
+}
+
+// indirection operator for vm::_ptr_base
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, T&> operator *(const vm::_ptr_base<T, AT>& ptr)
+{
+	return vm::get_ref<T>(vm::cast(ptr.m_addr));
+}
+
+// postfix increment operator for vm::_ptr_base
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>> operator ++(vm::_ptr_base<T, AT>& ptr, int)
+{
+	const AT result = ptr.m_addr;
+	ptr.m_addr += sizeof32(T);
+	return{ result };
+}
+
+// prefix increment operator for vm::_ptr_base
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>&> operator ++(vm::_ptr_base<T, AT>& ptr)
+{
+	ptr.m_addr += sizeof32(T);
+	return ptr;
+}
+
+// postfix decrement operator for vm::_ptr_base
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>> operator --(vm::_ptr_base<T, AT>& ptr, int)
+{
+	const AT result = ptr.m_addr;
+	ptr.m_addr -= sizeof32(T);
+	return{ result };
+}
+
+// prefix decrement operator for vm::_ptr_base
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>&> operator --(vm::_ptr_base<T, AT>& ptr)
+{
+	ptr.m_addr -= sizeof32(T);
+	return ptr;
+}
+
+// addition assignment operator for vm::_ptr_base (pointer += integer)
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>&> operator +=(vm::_ptr_base<T, AT>& ptr, to_ne_t<AT> count)
+{
+	ptr.m_addr += count * sizeof32(T);
+	return ptr;
+}
+
+// subtraction assignment operator for vm::_ptr_base (pointer -= integer)
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>&> operator -=(vm::_ptr_base<T, AT>& ptr, to_ne_t<AT> count)
+{
+	ptr.m_addr -= count * sizeof32(T);
+	return ptr;
+}
+
+// addition operator for vm::_ptr_base (pointer + integer)
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>> operator +(const vm::_ptr_base<T, AT>& ptr, to_ne_t<AT> count)
+{
+	return{ convert_le_be<AT>(ptr.m_addr + count * sizeof32(T)) };
+}
+
+// addition operator for vm::_ptr_base (integer + pointer)
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>> operator +(to_ne_t<AT> count, const vm::_ptr_base<T, AT>& ptr)
+{
+	return{ convert_le_be<AT>(ptr.m_addr + count * sizeof32(T)) };
+}
+
+// subtraction operator for vm::_ptr_base (pointer - integer)
+template<typename T, typename AT> inline vm::if_arithmetical_t<T, vm::_ptr_base<T, AT>> operator -(const vm::_ptr_base<T, AT>& ptr, to_ne_t<AT> count)
+{
+	return{ convert_le_be<AT>(ptr.m_addr - count * sizeof32(T)) };
+}
+
+// pointer difference operator for vm::_ptr_base
+template<typename T1, typename AT1, typename T2, typename AT2> inline std::enable_if_t<
+	!std::is_void<T1>::value &&
+	!std::is_void<T2>::value &&
+	!std::is_function<T1>::value &&
+	!std::is_function<T2>::value &&
+	std::is_same<std::remove_cv_t<T1>, std::remove_cv_t<T2>>::value,
+	u32> operator -(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return static_cast<u32>((left.m_addr - right.m_addr) / sizeof32(T1));
+}
+
+// comparison operator for vm::_ptr_base (pointer1 == pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator ==(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr == right.m_addr;
+}
+
+// comparison operator for vm::_ptr_base (pointer1 != pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator !=(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr != right.m_addr;
+}
+
+// comparison operator for vm::_ptr_base (pointer1 < pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator <(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr < right.m_addr;
+}
+
+// comparison operator for vm::_ptr_base (pointer1 <= pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator <=(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr <= right.m_addr;
+}
+
+// comparison operator for vm::_ptr_base (pointer1 > pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator >(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr > right.m_addr;
+}
+
+// comparison operator for vm::_ptr_base (pointer1 >= pointer2)
+template<typename T1, typename AT1, typename T2, typename AT2> vm::if_comparable_t<T1, T2, bool> operator >=(const vm::_ptr_base<T1, AT1>& left, const vm::_ptr_base<T2, AT2>& right)
+{
+	return left.m_addr >= right.m_addr;
+}
+
+// external specialization for is_be_t<> (true if AT is be_t<>)
+
+template<typename T, typename AT>
+struct is_be_t<vm::_ptr_base<T, AT>> : public std::integral_constant<bool, is_be_t<AT>::value>
+{
+};
+
+// external specialization for is_le_t<> (true if AT is le_t<>)
+
+template<typename T, typename AT>
+struct is_le_t<vm::_ptr_base<T, AT>> : public std::integral_constant<bool, is_le_t<AT>::value>
+{
+};
+
+// external specialization for to_ne_t<> (change AT endianness to native)
+
+template<typename T, typename AT>
+struct to_ne<vm::_ptr_base<T, AT>>
+{
+	using type = vm::_ptr_base<T, to_ne_t<AT>>;
+};
+
+// external specialization for to_be_t<> (change AT endianness to BE)
+
+template<typename T, typename AT>
+struct to_be<vm::_ptr_base<T, AT>>
+{
+	using type = vm::_ptr_base<T, to_be_t<AT>>;
+};
+
+// external specialization for to_le_t<> (change AT endianness to LE)
+
+template<typename T, typename AT>
+struct to_le<vm::_ptr_base<T, AT>>
+{
+	using type = vm::_ptr_base<T, to_le_t<AT>>;
+};
 
 namespace fmt
 {
 	// external specialization for fmt::format function
 
-	template<typename T, int lvl, typename AT>
-	struct unveil<vm::_ptr_base<T, lvl, AT>, false>
+	template<typename T, typename AT>
+	struct unveil<vm::_ptr_base<T, AT>, false>
 	{
-		typedef typename unveil<AT>::result_type result_type;
+		using result_type = typename unveil<AT>::result_type;
 
-		force_inline static result_type get_value(const vm::_ptr_base<T, lvl, AT>& arg)
+		force_inline static result_type get_value(const vm::_ptr_base<T, AT>& arg)
 		{
 			return unveil<AT>::get_value(arg.addr());
 		}
@@ -486,17 +391,17 @@ namespace fmt
 template<typename T, bool is_enum>
 struct cast_ppu_gpr;
 
-template<typename T, int lvl, typename AT>
-struct cast_ppu_gpr<vm::_ptr_base<T, lvl, AT>, false>
+template<typename T, typename AT>
+struct cast_ppu_gpr<vm::_ptr_base<T, AT>, false>
 {
-	force_inline static u64 to_gpr(const vm::_ptr_base<T, lvl, AT>& value)
+	force_inline static u64 to_gpr(const vm::_ptr_base<T, AT>& value)
 	{
 		return cast_ppu_gpr<AT, std::is_enum<AT>::value>::to_gpr(value.addr());
 	}
 
-	force_inline static vm::_ptr_base<T, lvl, AT> from_gpr(const u64 reg)
+	force_inline static vm::_ptr_base<T, AT> from_gpr(const u64 reg)
 	{
-		return vm::_ptr_base<T, lvl, AT>::make(cast_ppu_gpr<AT, std::is_enum<AT>::value>::from_gpr(reg));
+		return vm::_ptr_base<T, AT>::make(cast_ppu_gpr<AT, std::is_enum<AT>::value>::from_gpr(reg));
 	}
 };
 
@@ -505,16 +410,16 @@ struct cast_ppu_gpr<vm::_ptr_base<T, lvl, AT>, false>
 template<typename T, bool is_enum>
 struct cast_armv7_gpr;
 
-template<typename T, int lvl, typename AT>
-struct cast_armv7_gpr<vm::_ptr_base<T, lvl, AT>, false>
+template<typename T, typename AT>
+struct cast_armv7_gpr<vm::_ptr_base<T, AT>, false>
 {
-	force_inline static u32 to_gpr(const vm::_ptr_base<T, lvl, AT>& value)
+	force_inline static u32 to_gpr(const vm::_ptr_base<T, AT>& value)
 	{
 		return cast_armv7_gpr<AT, std::is_enum<AT>::value>::to_gpr(value.addr());
 	}
 
-	force_inline static vm::_ptr_base<T, lvl, AT> from_gpr(const u32 reg)
+	force_inline static vm::_ptr_base<T, AT> from_gpr(const u32 reg)
 	{
-		return vm::_ptr_base<T, lvl, AT>::make(cast_armv7_gpr<AT, std::is_enum<AT>::value>::from_gpr(reg));
+		return vm::_ptr_base<T, AT>::make(cast_armv7_gpr<AT, std::is_enum<AT>::value>::from_gpr(reg));
 	}
 };
