@@ -3,80 +3,88 @@
 namespace vm
 {
 	template<typename T, typename AT = u32>
-	class _ref_base
+	struct _ref_base
 	{
-	protected:
 		AT m_addr;
 
-	public:
-		typedef T type;
 		static_assert(!std::is_pointer<T>::value, "vm::_ref_base<> error: invalid type (pointer)");
 		static_assert(!std::is_reference<T>::value, "vm::_ref_base<> error: invalid type (reference)");
-		typedef typename remove_be_t<T>::type le_type;
-		typedef typename to_be_t<T>::type be_type;
-
-		operator T&()
-		{
-			return get_ref<T>(m_addr);
-		}
-
-		operator const T&() const
-		{
-			return get_ref<const T>(m_addr);
-		}
+		static_assert(!std::is_function<T>::value, "vm::_ref_base<> error: invalid type (function)");
+		static_assert(!std::is_void<T>::value, "vm::_ref_base<> error: invalid type (void)");
 
 		AT addr() const
 		{
 			return m_addr;
 		}
 
-		static _ref_base make(const AT& addr)
+		template<typename AT2 = AT> static _ref_base make(const AT2& addr)
 		{
-			return reinterpret_cast<_ref_base&>(addr);
+			return{ convert_le_be<AT>(addr) };
 		}
 
-		_ref_base& operator = (le_type right)
+		T& get_ref() const
 		{
-			get_ref<T>(m_addr) = right;
-			return *this;
+			return vm::get_ref<T>(vm::cast(m_addr));
 		}
 
-		const _ref_base& operator = (le_type right) const
+		T& priv_ref() const
 		{
-			get_ref<T>(m_addr) = right;
-			return *this;
+			return vm::priv_ref<T>(vm::cast(m_addr));
 		}
 
-		_ref_base& operator = (be_type right)
+		// TODO: conversion operator (seems hard to define it correctly)
+		//template<typename CT, typename dummy = std::enable_if_t<std::is_convertible<T, CT>::value || std::is_convertible<to_ne_t<T>, CT>::value>> operator CT() const
+		//{
+		//	return get_ref();
+		//}
+
+		operator to_ne_t<T>() const
 		{
-			get_ref<T>(m_addr) = right;
-			return *this;
+			return get_ref();
 		}
 
-		const _ref_base& operator = (be_type right) const
+		explicit operator T&() const
 		{
-			get_ref<T>(m_addr) = right;
-			return *this;
+			return get_ref();
+		}
+
+		// copy assignment operator:
+		// returns T& by default, this may be wrong if called assignment operator has different return type
+		T& operator =(const _ref_base& right)
+		{
+			static_assert(!std::is_const<T>::value, "vm::_ref_base<> error: operator= is not available for const reference");
+
+			return get_ref() = right.get_ref();
+		}
+
+		template<typename CT, typename AT2> auto operator =(const _ref_base<CT, AT2>& right) const -> decltype(std::declval<T&>() = std::declval<CT>())
+		{
+			return get_ref() = right.get_ref();
+		}
+
+		template<typename CT> auto operator =(const CT& right) const -> decltype(std::declval<T&>() = std::declval<CT>())
+		{
+			return get_ref() = right;
 		}
 	};
 
 	// Native endianness reference to LE data
-	template<typename T, typename AT = u32> using refl = _ref_base<typename to_le_t<T>::type, AT>;
+	template<typename T, typename AT = u32> using refl = _ref_base<to_le_t<T>, AT>;
 
 	// Native endianness reference to BE data
-	template<typename T, typename AT = u32> using refb = _ref_base<typename to_be_t<T>::type, AT>;
+	template<typename T, typename AT = u32> using refb = _ref_base<to_be_t<T>, AT>;
 
 	// BE reference to LE data
-	template<typename T, typename AT = u32> using brefl = _ref_base<typename to_le_t<T>::type, typename to_be_t<AT>::type>;
+	template<typename T, typename AT = u32> using brefl = _ref_base<to_le_t<T>, to_be_t<AT>>;
 
 	// BE reference to BE data
-	template<typename T, typename AT = u32> using brefb = _ref_base<typename to_be_t<T>::type, typename to_be_t<AT>::type>;
+	template<typename T, typename AT = u32> using brefb = _ref_base<to_be_t<T>, to_be_t<AT>>;
 
 	// LE reference to LE data
-	template<typename T, typename AT = u32> using lrefl = _ref_base<typename to_le_t<T>::type, typename to_le_t<AT>::type>;
+	template<typename T, typename AT = u32> using lrefl = _ref_base<to_le_t<T>, to_le_t<AT>>;
 
 	// LE reference to BE data
-	template<typename T, typename AT = u32> using lrefb = _ref_base<typename to_be_t<T>::type, typename to_le_t<AT>::type>;
+	template<typename T, typename AT = u32> using lrefb = _ref_base<to_be_t<T>, to_le_t<AT>>;
 
 	namespace ps3
 	{
@@ -90,15 +98,137 @@ namespace vm
 	namespace psv
 	{
 		// default reference for PSV HLE functions (Native endianness reference to LE data)
-		template<typename T, typename AT = u32> using ref = refl<T, AT>;
+		template<typename T> using ref = refl<T>;
 
 		// default reference for PSV HLE structures (LE reference to LE data)
-		template<typename T, typename AT = u32> using lref = lrefl<T, AT>;
+		template<typename T> using lref = lrefl<T>;
 	}
 
 	//PS3 emulation is main now, so lets it be as default
 	using namespace ps3;
 }
+
+// postfix increment operator for vm::_ref_base
+template<typename T, typename AT> inline auto operator ++(const vm::_ref_base<T, AT>& ref, int) -> decltype(std::declval<T&>()++)
+{
+	return ref.get_ref()++;
+}
+
+// prefix increment operator for vm::_ref_base
+template<typename T, typename AT> inline auto operator ++(const vm::_ref_base<T, AT>& ref) -> decltype(++std::declval<T&>())
+{
+	return ++ref.get_ref();
+}
+
+// postfix decrement operator for vm::_ref_base
+template<typename T, typename AT> inline auto operator --(const vm::_ref_base<T, AT>& ref, int) -> decltype(std::declval<T&>()--)
+{
+	return ref.get_ref()--;
+}
+
+// prefix decrement operator for vm::_ref_base
+template<typename T, typename AT> inline auto operator --(const vm::_ref_base<T, AT>& ref) -> decltype(--std::declval<T&>())
+{
+	return --ref.get_ref();
+}
+
+// addition assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator +=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() += std::declval<T2>())
+{
+	return ref.get_ref() += right;
+}
+
+// subtraction assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator -=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() -= std::declval<T2>())
+{
+	return ref.get_ref() -= right;
+}
+
+// multiplication assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator *=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() *= std::declval<T2>())
+{
+	return ref.get_ref() *= right;
+}
+
+// division assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator /=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() /= std::declval<T2>())
+{
+	return ref.get_ref() /= right;
+}
+
+// modulo assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator %=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() %= std::declval<T2>())
+{
+	return ref.get_ref() %= right;
+}
+
+// bitwise AND assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator &=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() &= std::declval<T2>())
+{
+	return ref.get_ref() &= right;
+}
+
+// bitwise OR assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator |=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() |= std::declval<T2>())
+{
+	return ref.get_ref() |= right;
+}
+
+// bitwise XOR assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator ^=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() ^= std::declval<T2>())
+{
+	return ref.get_ref() ^= right;
+}
+
+// bitwise left shift assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator <<=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() <<= std::declval<T2>())
+{
+	return ref.get_ref() <<= right;
+}
+
+// bitwise right shift assignment operator for vm::_ref_base
+template<typename T, typename AT, typename T2> inline auto operator >>=(const vm::_ref_base<T, AT>& ref, const T2& right) -> decltype(std::declval<T&>() >>= std::declval<T2>())
+{
+	return ref.get_ref() >>= right;
+}
+
+// external specialization for is_be_t<> (true if AT's endianness is BE)
+
+template<typename T, typename AT>
+struct is_be_t<vm::_ref_base<T, AT>> : public std::integral_constant<bool, is_be_t<AT>::value>
+{
+};
+
+// external specialization for is_le_t<> (true if AT's endianness is LE)
+
+template<typename T, typename AT>
+struct is_le_t<vm::_ref_base<T, AT>> : public std::integral_constant<bool, is_le_t<AT>::value>
+{
+};
+
+// external specialization for to_ne_t<> (change AT's endianness to native)
+
+template<typename T, typename AT>
+struct to_ne<vm::_ref_base<T, AT>>
+{
+	using type = vm::_ref_base<T, to_ne_t<AT>>;
+};
+
+// external specialization for to_be_t<> (change AT's endianness to BE)
+
+template<typename T, typename AT>
+struct to_be<vm::_ref_base<T, AT>>
+{
+	using type = vm::_ref_base<T, to_be_t<AT>>;
+};
+
+// external specialization for to_le_t<> (change AT's endianness to LE)
+
+template<typename T, typename AT>
+struct to_le<vm::_ref_base<T, AT>>
+{
+	using type = vm::_ref_base<T, to_le_t<AT>>;
+};
 
 namespace fmt
 {
@@ -107,7 +237,7 @@ namespace fmt
 	template<typename T, typename AT>
 	struct unveil<vm::_ref_base<T, AT>, false>
 	{
-		typedef typename unveil<AT>::result_type result_type;
+		using result_type = typename unveil<AT>::result_type;
 
 		force_inline static result_type get_value(const vm::_ref_base<T, AT>& arg)
 		{
