@@ -108,8 +108,7 @@ enum
 };
 
 // Request handle for clan API
-struct SceNpClansRequest;
-typedef vm::ptr<SceNpClansRequest> SceNpClansRequestHandle;
+using SceNpClansRequestHandle = vm::ptr<struct SceNpClansRequest>;
 
 // Paging request structure
 struct SceNpClansPagingRequest
