@@ -5,7 +5,7 @@
 namespace vm
 {
 	template<typename AT, typename RT, typename... T>
-	force_inline RT _ptr_base<RT(T...), 1, AT>::operator()(ARMv7Context& context, T... args) const
+	force_inline RT _ptr_base<RT(T...), AT>::operator()(ARMv7Context& context, T... args) const
 	{
 		return psv_func_detail::func_caller<RT, T...>::call(context, vm::cast(this->addr()), args...);
 	}
