@@ -301,7 +301,7 @@ public:
 
 				m_arb_shader += fmt::format("#%d ", i) + param_type + param_name + param_semantic + param_const + "\n";
 
-				offset += sizeof(CgBinaryParameter);
+				offset += sizeof32(CgBinaryParameter);
 			}
 
 			m_arb_shader += "\n";
@@ -355,7 +355,7 @@ public:
 
 				m_arb_shader += fmt::format("#%d ", i) + param_type + param_name + param_semantic + param_const + "\n";
 
-				offset += sizeof(CgBinaryParameter);
+				offset += sizeof32(CgBinaryParameter);
 			}
 
 			m_arb_shader += "\n";
