@@ -1,6 +1,6 @@
 #pragma once
 
-//Return Codes
+// Return Codes
 enum
 {
 	CELL_GIFDEC_ERROR_OPEN_FILE     = 0x80611300,
@@ -13,29 +13,94 @@ enum
 	CELL_GIFDEC_ERROR_CB_PARAM      = 0x80611307,
 };
 
-enum CellGifDecStreamSrcSel
+enum CellGifDecStreamSrcSel : s32
 {
 	CELL_GIFDEC_FILE   = 0, // Input from a file
 	CELL_GIFDEC_BUFFER = 1, // Input from a buffer
 };
 
-enum CellGifDecColorSpace
+enum CellGifDecSpuThreadEna : s32
 {
-	CELL_GIFDEC_RGBA = 10,
-	CELL_GIFDEC_ARGB = 20,
+	CELL_GIFDEC_SPU_THREAD_DISABLE = 0, // Do not use SPU threads
+	CELL_GIFDEC_SPU_THREAD_ENABLE  = 1, // Use SPU threads
 };
 
-enum CellGifDecRecordType
+enum CellGifDecRecordType : s32
 {
 	CELL_GIFDEC_RECORD_TYPE_IMAGE_DESC = 1, // Image data block
 	CELL_GIFDEC_RECORD_TYPE_EXTENSION  = 2, // Extension block
 	CELL_GIFDEC_RECORD_TYPE_TERMINATE  = 3, // Trailer block
 };
 
-enum CellGifDecDecodeStatus
+enum CellGifDecColorSpace : s32
 {
-	CELL_GIFDEC_DEC_STATUS_FINISH = 0, //Decoding finished
-	CELL_GIFDEC_DEC_STATUS_STOP   = 1, //Decoding halted
+	CELL_GIFDEC_RGBA = 10, // RGBA
+	CELL_GIFDEC_ARGB = 20, // ARGB
+};
+
+enum CellGifDecCommand : s32
+{
+	CELL_GIFDEC_CONTINUE = 0, // Continue decoding
+	CELL_GIFDEC_STOP     = 1, // Force decoding to stop
+};
+
+enum CellGifDecDecodeStatus : s32
+{
+	CELL_GIFDEC_DEC_STATUS_FINISH = 0, // Decoding finished
+	CELL_GIFDEC_DEC_STATUS_STOP   = 1, // Decoding was stopped
+};
+
+// Handles
+using CellGifDecMainHandle = vm::ptr<struct GifDecoder>;
+using CellGifDecSubHandle = u32; // vm::ptr<struct GifStream>;
+
+// Callbacks for memory management
+using CellGifDecCbControlMalloc = func_def<vm::ptr<void>(u32 size, vm::ptr<void> cbCtrlMallocArg)>;
+using CellGifDecCbControlFree = func_def<s32(vm::ptr<void> ptr, vm::ptr<void> cbCtrlFreeArg)>;
+
+// Structs
+struct CellGifDecThreadInParam
+{
+	be_t<s32> spuThreadEnable; // CellGifDecSpuThreadEna
+	be_t<u32> ppuThreadPriority;
+	be_t<u32> spuThreadPriority;
+	vm::bptr<CellGifDecCbControlMalloc> cbCtrlMallocFunc;
+	vm::bptr<void> cbCtrlMallocArg;
+	vm::bptr<CellGifDecCbControlFree> cbCtrlFreeFunc;
+	vm::bptr<void> cbCtrlFreeArg;
+};
+
+struct CellGifDecThreadOutParam
+{
+	be_t<u32> gifCodecVersion;
+};
+
+struct CellGifDecExtThreadInParam
+{
+	vm::bptr<struct CellSpurs> spurs;
+	u8 priority[8];
+	be_t<u32> maxContention;
+};
+
+struct CellGifDecExtThreadOutParam
+{
+	be_t<u64> reserved;
+};
+
+struct CellGifDecSrc
+{
+	be_t<s32> srcSelect; // CellGifDecStreamSrcSel
+	vm::bptr<const char> fileName;
+	be_t<s64> fileOffset;
+	be_t<u32> fileSize;
+	vm::bptr<void> streamPtr;
+	be_t<u32> streamSize;
+	be_t<s32> spuThreadEnable; // CellGifDecSpuThreadEna
+};
+
+struct CellGifDecOpnInfo
+{
+	be_t<u32> initSpaceAllocated;
 };
 
 struct CellGifDecInfo
@@ -50,24 +115,13 @@ struct CellGifDecInfo
 	be_t<u32> SPixelAspectRatio;
 };
 
-struct CellGifDecSrc
-{
-	be_t<u32> srcSelect;
-	vm::bptr<const char> fileName;
-	be_t<s64> fileOffset;
-	be_t<u64> fileSize;
-	be_t<u32> streamPtr;
-	be_t<u32> streamSize;
-	be_t<u32> spuThreadEnable;
-};
-
 struct CellGifDecInParam
 {
-	be_t<u32> commandPtr;
-	be_t<u32> colorSpace; // CellGifDecColorSpace
-	be_t<u8> outputColorAlpha1;
-	be_t<u8> outputColorAlpha2;
-	be_t<u8> reserved[2];
+	vm::bptr<volatile s32> commandPtr; // CellGifDecCommand
+	be_t<s32> colorSpace; // CellGifDecColorSpace
+	u8 outputColorAlpha1;
+	u8 outputColorAlpha2;
+	u8 reserved[2];
 };
 
 struct CellGifDecOutParam
@@ -77,26 +131,21 @@ struct CellGifDecOutParam
 	be_t<u32> outputHeight;
 	be_t<u32> outputComponents;
 	be_t<u32> outputBitDepth;
-	be_t<u32> outputColorSpace;	// CellGifDecColorSpace
+	be_t<s32> outputColorSpace; // CellGifDecColorSpace
 	be_t<u32> useMemorySpace;
 };
 
 struct CellGifDecExtension
 {
-	be_t<u8> label;
-	be_t<u32> data;
+	u8 label;
+	vm::ptr<u8> data;
 };
 
 struct CellGifDecDataOutInfo
 {
-	be_t<u32> recordType;
+	be_t<s32> recordType; // CellGifDecRecordType
 	CellGifDecExtension outExtension;
-	be_t<u32> status;
-};
-
-struct CellGifDecOpnInfo
-{
-	be_t<u32> initSpaceAllocated;
+	be_t<s32> status; // CellGifDecDecodeStatus
 };
 
 struct CellGifDecDataCtrlParam
@@ -104,8 +153,12 @@ struct CellGifDecDataCtrlParam
 	be_t<u64> outputBytesPerLine;
 };
 
-//Custom structs
-struct CellGifDecSubHandle
+// Custom structs
+struct GifDecoder
+{
+};
+
+struct GifStream
 {
 	u32 fd;
 	u64 fileSize;
